feat: extend getConfiguration test to check the rate limiter configuration
