+++ .github/workflows/cypress-ete.yml
@@ -10,9 +10,7 @@ jobs:
     name: Cypress (ete)
     runs-on: ubuntu-latest
     services:
-      # Label used to access the service container
       redis:
-        # Docker Hub image
         image: redis
         # Set health checks to wait until redis has started
         options: >-
@@ -21,7 +19,7 @@ jobs:
           --health-timeout 5s
           --health-retries 5
         ports:
-          # Maps port 6379 on service container to the host
+          # Maps Redis port 6379 on service container to the host
           - 6379:6379
     steps:
       - name: Checkout
