+++ .github/workflows/cypress-mocked.yml
@@ -46,6 +46,7 @@ jobs:
           SENTRY_DSN: ''
           REDIS_HOST: localhost
           REDIS_PASSWORD: ${{ secrets.REDIS_PASSWORD }}
+          RATE_LIMITER_CONFIG: '{"enabled":false,"defaultBuckets":{"globalBucket":{"capacity":500,"addTokenMs":50}}'
         with:
           build: yarn build
           start: yarn start, yarn mock-server
