+++ src/server/lib/__tests__/getConfiguration.test.ts
@@ -1,7 +1,5 @@
 /* eslint-disable functional/immutable-data */
 /* Linting rule disable as unit test needs to mutate env */
-
-import { getConfiguration } from '@/server/lib/getConfiguration';
 import type { Configuration } from '@/server/models/Configuration';
 import {
   GA_UID,
@@ -21,7 +19,24 @@ describe('getConfiguration', () => {
     process.env = ORIGINAL_ENVIRONMENT_VARIABLES;
   });
 
-  test('it returns the configuration object with the correct values', () => {
+  const rateLimiterConfig = `
+  {
+    "enabled": true,
+    "defaultBuckets": {
+      "globalBucket": { "capacity": 500, "addTokenMs": 50 },
+      "ipBucket": { "capacity": 100, "addTokenMs": 50 },
+      "emailBucket": { "capacity": 100, "addTokenMs": 50 },
+      "oktaIdentifierBucket": { "capacity": 100, "addTokenMs": 50 },
+      "accessTokenBucket": { "capacity": 100, "addTokenMs": 50 }
+    },
+    "routeBuckets": {
+      "/signin": {
+        "globalBucket": { "capacity": 500, "addTokenMs": 50 }
+      }
+    }
+  }`;
+
+  test('it returns the configuration object with the correct values', async () => {
     process.env.PORT = '9000';
     process.env.IDAPI_CLIENT_ACCESS_TOKEN = 'idapi_api_key';
     process.env.IDAPI_BASE_URL = 'http://localhost:1234';
@@ -45,24 +60,9 @@ describe('getConfiguration', () => {
     process.env.GITHUB_RUN_NUMBER = '5';
     process.env.REDIS_PASSWORD = 'redispassword';
     process.env.REDIS_HOST = 'localhost:1234';
+    process.env.RATE_LIMITER_CONFIG = rateLimiterConfig;
 
-    process.env.RATE_LIMIT_CONFIGURATION = `
-    {
-      "enabled": true,
-      "defaultBuckets": {
-        "globalBucket": { "capacity": 500, "addTokenMs": 50 },
-        "ipBucket": { "capacity": 100, "addTokenMs": 50 },
-        "emailBucket": { "capacity": 100, "addTokenMs": 50 },
-        "oktaIdentifierBucket": { "capacity": 100, "addTokenMs": 50 },
-        "accessTokenBucket": { "capacity": 100, "addTokenMs": 50 }
-      },
-      "routeBuckets": {
-        "/signin": {
-          "globalBucket": { "capacity": 500, "addTokenMs": 50 }
-        }
-      }
-    }`;
-
+    const { getConfiguration } = await import('@/server/lib/getConfiguration');
     const output = getConfiguration();
     const expected = {
       port: 9000,
@@ -127,8 +127,22 @@ describe('getConfiguration', () => {
     expect(output).toEqual(expected);
   });
 
-  test('it throws and exception if the port is not set', () => {
-    process.env = {};
+  test('it throws an exception if the port is not set', async () => {
+    process.env = {
+      RATE_LIMITER_CONFIG: rateLimiterConfig,
+    };
+    const { getConfiguration } = await import('@/server/lib/getConfiguration');
     expect(() => getConfiguration()).toThrow();
+  });
+
+  test('it throws an exception if the rate limiter configuration is not set', async () => {
+    process.env = {};
+    try {
+      await import('@/server/lib/getConfiguration');
+    } catch (error) {
+      expect(error).toEqual(
+        Error('Rate limiter configuration missing or malformed'),
+      );
+    }
   });
 });
